requirements.txt
absl-py==2.1.0
grpcio==1.66.2
grpcio_tools==1.66.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-grpc==0.48b0
